Separated ingester logic to new service
