+++ services/hype-backend/src/modules/group/group.module.ts
@@ -1,9 +1,11 @@
-import { Module, OnModuleInit } from '@nestjs/common';
+import { DynamicModule, Module, OnModuleInit, Provider } from '@nestjs/common';
+import { ModuleRef } from '@nestjs/core';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { GroupService } from './group.service';
 import { GroupController } from './group.controller';
 import { Group } from '../../entities';
 import { IpfsModule } from '../ipfs';
+import { IngesterService } from './ingester.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Group]), IpfsModule],
@@ -11,9 +13,27 @@ import { IpfsModule } from '../ipfs';
   providers: [GroupService],
 })
 export class GroupModule implements OnModuleInit {
-  constructor(private readonly groupService: GroupService) {}
+  static type = 'web';
+  static forRoot(type = 'web'): DynamicModule {
+    GroupModule.type = type;
+    let providers: Provider[] = [];
+    if (type === 'cron') {
+      providers = [IngesterService];
+    }
+    return {
+      module: GroupModule,
+      providers,
+    };
+  }
+
+  constructor(private readonly moduleRef: ModuleRef) {}
 
   onModuleInit() {
-    this.groupService.collectDataForIngesters();
+    if (GroupModule.type === 'cron') {
+      const ingesterService = this.moduleRef.get(IngesterService, {
+        strict: false,
+      });
+      ingesterService?.collectDataForIngesters();
+    }
   }
 }
